Update method to disable X11 screensaver
On modern systems `xdg-screensaver reset` does not necessarily
reset the timer. It depends on DPMS, and according to spec turns off
the screensaver (unlock the screen), via `xset dpms force on`.

Explicit window suspend/resume make use of DBus
org.freedesktop.ScreenSaver.Inhibit, which is not be supported
but all systems, so we fall back to good old `XResetScreenSaver`
which does reset the activation timer.