When dragging automation points, only drag horizontally or vertically
The most common drag operation on an automation point is with the intent
of either moving it in time or in value. That was very hard to achieve
when both dimensions were free.

Slightly similar to "Ignore Y-axis when adding new automation-points",
we thus "snap" the move to either being horizontal or vertical.

In the rare occation where we want to move in both dimensions at once,
we can still and fully intuitively do that with 2 drag operations.