Refactor to let Editor::can_remove_control_point take a ControlPoint reference
Avoid retrieving control_point from Item both in
Editor::remove_control_point and in Editor::can_remove_control_point .

remove_control_point is called from Editor::button_release_handler, and
is very view-ish, so it is fine that it takes an Item.

can_remove_control_point is slightly more low level and model-ish
(almost so much that it could live in AudioRegionGainLine), so it kind
of makes sense that it works on the ControlPoint.

There is no extra code. The retrieval of ControlPoint moves from
can_remove_control_point to Editor::popup_control_point_context_menu .

This refactoring will make it easier to implement deletion of multiple
selected points.