Fall back to using plain ffmpeg and ffprobe tools if not finding the harvid versions
On Windows, it will look for system installed ffmpeg - for example in
C:\Program Files\ffmpeg\ffmpeg.exe . On Linux/Mac it only looks for
ffmeg_harvid and ffprobe_harvid in $PATH.

https://manual.ardour.org/video-timeline/setup/ suggests creating
symlinks from the plain ffmpeg tools to the harvid name. But in my
opinion: The _harvid suffix is great for the usecase of shipping
self-contained binaries. But it seems a bit unfortunate that local or
distribution builds of Ardour has to put an extra file in $PATH just
because Ardour wants to use it. That doesn't scale.

Instead, look for plain ffmpeg in $PATH as fallback if no harvid version
is found.

This will enable MP3 export on Fedora with the defacto standard ffmpeg
from rpmfusion. It also seems like it would fix for example
https://bugs.gentoo.org/601840 .