Allow removing multiple control points
It was surprising that removing on multiple selected control points only
removed one of them.

This allows selecting multiple points by dragging a box or
Ctrl-clicking, then removing all from the context menu of one of the
points.

Some more trivial implementations of this suffered from problems when
deletion of control points in the model also invalidated the iterator
over selected control points.